# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart Market Watchlist REST API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface RegisterInput {
  /**
     * @minLength 2
     * @maxLength 80
     */
  name: string;
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}

export interface LoginInput {
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}

export interface WatchlistItem {
  id: string;
  symbol: string;
  companyName: string;
  exchange: string;
  thresholdPercent: number;
  /** @nullable */
  lastViewedAt?: string | null;
  /** @nullable */
  lastCheckpointPrice?: number | null;
  /** @nullable */
  lastCheckpointTimestamp?: string | null;
}

export interface Watchlist {
  id: string;
  name: string;
  items: WatchlistItem[];
}

export interface WatchlistItemInput {
  /** @pattern ^[A-Za-z0-9._-]{1,12}$ */
  symbol: string;
}

export interface WatchlistSettingsInput {
  /**
     * @minimum 1
     * @maximum 25
     */
  thresholdPercent: number;
}

export interface StockSearchResult {
  symbol: string;
  companyName: string;
  exchange: string;
  sector: string;
}

export interface Signal {
  type: string;
  label: string;
  value: number;
  severity: string;
  contribution: number;
  available?: boolean;
}

export interface Explanation {
  fact: string;
  signal: string;
  context: string;
  explanation: string;
  unknown: string;
  source: string;
}

export interface Attention {
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  severity: string;
  sinceLastCheckPercent: number;
  /** @nullable */
  sinceLastCheckAt: string | null;
  signals: Signal[];
  explanation: Explanation;
}

export interface MarketItem {
  symbol: string;
  companyName: string;
  exchange: string;
  sector?: string;
  price: number;
  changePercent: number;
  /** @nullable */
  volumeRatio: number | null;
  sectorChangePercent: number;
  marketChangePercent: number;
  observedAt: string;
  source: string;
  isStale: boolean;
  marketState?: string;
  attention: Attention;
  /** @nullable */
  activeEventId?: string | null;
}

export interface MarketWatchlistResponse {
  marketState: string;
  updatedAt: string;
  items: MarketItem[];
  attentionItems: MarketItem[];
}

export interface TimelineEntry {
  timestamp: string;
  label: string;
  detail: string;
  kind: string;
}

export interface Checkpoint {
  /** @nullable */
  lastViewedAt: string | null;
  /** @nullable */
  lastCheckpointPrice: number | null;
  /** @nullable */
  lastCheckpointTimestamp: string | null;
}

export interface StockDetail {
  item: MarketItem;
  checkpoint: Checkpoint;
  timeline: TimelineEntry[];
}

export interface Event {
  id: string;
  symbol: string;
  companyName: string;
  detectedAt: string;
  score: number;
  severity: string;
  status: string;
  /** @nullable */
  acknowledgedAt?: string | null;
  explanation: Explanation;
  signals: Signal[];
}

export type SearchStocksParams = {
/**
 * @minLength 1
 * @maxLength 40
 */
q: string;
};