# lib/api-zod/src/generated/types/checkpoint.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart Market Watchlist REST API
 * OpenAPI spec version: 0.1.0
 */

export interface Checkpoint {
  /** @nullable */
  lastViewedAt: Date | null;
  /** @nullable */
  lastCheckpointPrice: number | null;
  /** @nullable */
  lastCheckpointTimestamp: Date | null;
}