# lib/api-zod/src/generated/types/explanation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart Market Watchlist REST API
 * OpenAPI spec version: 0.1.0
 */

export interface Explanation {
  fact: string;
  signal: string;
  context: string;
  explanation: string;
  unknown: string;
  source: string;
}