# lib/api-zod/src/generated/types/marketWatchlistResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart Market Watchlist REST API
 * OpenAPI spec version: 0.1.0
 */
import type { MarketItem } from './marketItem';

export interface MarketWatchlistResponse {
  marketState: string;
  updatedAt: Date;
  items: MarketItem[];
  attentionItems: MarketItem[];
}