# lib/api-zod/src/generated/types/watchlist.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart Market Watchlist REST API
 * OpenAPI spec version: 0.1.0
 */
import type { WatchlistItem } from './watchlistItem';

export interface Watchlist {
  id: string;
  name: string;
  items: WatchlistItem[];
}