# lib/api-zod/src/generated/types/watchlistSettingsInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart Market Watchlist REST API
 * OpenAPI spec version: 0.1.0
 */

export interface WatchlistSettingsInput {
  /**
     * @minimum 1
     * @maximum 25
     */
  thresholdPercent: number;
}